Remove warehouseId from request, derive warehouse, unify IDs to Long

+++ agape-api/src/main/java/hr/agape/document/dto/DocumentDescriptorResponseDTO.java
@@ -14,9 +14,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class DocumentDescriptorResponseDTO {
-    private Integer documentId;     // SD_SIFREG.DOKUMENT_ID
-    private String  documentCode;   // SD_SIFREZ.DOKUMENTID (e.g. OTPREMNICA)
-    private String  displayName;    // SD_SIFREZ.NAZIVDOKUMENTA (e.g. IZDATNICA)
-    private Integer inOutFlag;      // SD_SIFREZ.ULAZIZLAZ (1=IN,4=OUT,…)
-    private Integer changesStock;   // SD_SIFREZ.MIJENJAZALIHU (1=yes,0=no)
+    private Long documentId;    // SD_SIFREG.DOKUMENT_ID
+    private String documentCode;  // SD_SIFREZ.DOKUMENTID (e.g. OTPREMNICA)
+    private String displayName;   // SD_SIFREZ.NAZIVDOKUMENTA (e.g. IZDATNICA)
+    private Long inOutFlag;     // SD_SIFREZ.ULAZIZLAZ (1=IN,4=OUT,…)
+    private Long changesStock;  // SD_SIFREZ.MIJENJAZALIHU (1=yes,0=no)
 }
